refactored while-loop in client(app/main.py

+++ client/app/main.py
@@ -31,31 +31,44 @@ if len(pads) == 0:
 
 # Cache for outputs from gamepad
 output_dict = {"BTN_TL": 0, "BTN_TR": 0, "ABS_RX": 0, "ABS_Y": 0, "BTN_EAST": 0}
+autopilot_active = False
 
 # Start
 time.sleep(config.START_SLEEP_TIME)
 logging.info("Main process is ready!")
+
 try:
     while True:
 
-        # Read inputs from gamepad
+        # Read all inputs from gamepad
         events = inputs.get_gamepad()
+
+        # autopilot
+        if len(events) == 0:
+            if output_dict["BTN_EAST"] == 1:
+                output_dict = autopilot.predict(output_dict)
+                driving.drive(output_dict)
+                autopilot_active = True
+            if (output_dict["BTN_EAST"] == 0) and autopilot_active:
+                output_dict["ABS_RX"] = 0
+                output_dict["ABS_Y"] = 0
+                autopilot_active = False
+            continue
 
         for event in events:
 
             # Check if shutdown is requested
             if event.code == "BTN_MODE":
+                driving.disable()
+                time.sleep(config.MAIN_SLEEP_TIME)
+                GPIO.cleanup()
                 call("sudo poweroff", shell=True)
 
+            # update output_dict
             if event.code in output_dict.keys():
-
-                # update Cache
-                output_dict[event.code] = event.state
-
-                if output_dict["BTN_EAST"] == 1:  # RED Button
-                    output_dict = autopilot.predict(output_dict)
-                driving.drive(output_dict)
-                db.set(config.GAMEPAD, json.dumps(output_dict))  # update_cache
+                output_dict[event.code] = event.state  # update
+                driving.drive(output_dict)  # drive
+                db.set(config.GAMEPAD, json.dumps(output_dict))  # update redis cache
 
 except KeyboardInterrupt:
     driving.disable()
